responsible-ai-telemetry/requirements.txt
annotated-types   
anyio             
blinker          
certifi           
click             
colorama         
dnspython         
elastic-transport==8.4.1
elasticsearch==8.10.0
elasticsearch-dsl==8.9.0
fastapi==0.115.12
Flask==3.0.0
h11
idna              
itsdangerous      
Jinja2            
MarkupSafe       
pip               
pydantic          
pydantic_core     
pymongo==4.6.3
python-dateutil   
python-dotenv     
pytz              
PyYAML            
setuptools        
six              
sniffio          
starlette==0.40.0   
typing_extensions 
tzdata            
urllib3          
uvicorn           
Werkzeug          